discord_verify/requirements.txt
boto3==1.14.9
Cerberus==1.3.2
pymongo==3.10.1
discord.py==1.3.4
requests==2.24.0
